refactor: move to tasks router + use pydantic model
