+++ src/backend/app/projects/project_crud.py
@@ -24,7 +24,6 @@ import time
 import uuid
 from asyncio import gather
 from concurrent.futures import ThreadPoolExecutor, wait
-from datetime import datetime, timedelta
 from importlib.resources import files as pkg_files
 from io import BytesIO
 from typing import List, Optional, Union
@@ -2459,52 +2458,3 @@ def count_user_contributions(db: Session, user_id: int, project_id: int) -> int:
     )
 
     return contributions_count
-
-
-async def count_validated_and_mapped_tasks(
-    task_history: db_models.DbTaskHistory, end_date: datetime, db: Session
-):
-    """Counts the number of validated and mapped tasks.
-
-    Args:
-        task_history: The task history records to count.
-        end_date: The end date of the date range.
-        db: The database session.
-
-    Returns:
-        A list of dictionaries with following keys:
-        - 'date': The date in the format 'MM/DD'.
-        - 'validated': The cumulative count of validated tasks.
-        - 'mapped': The cumulative count of mapped tasks.
-    """
-    cumulative_counts = {}
-    results = []
-
-    current_date = end_date
-    while current_date <= datetime.now():
-        date_str = current_date.strftime("%m/%d")
-        cumulative_counts = {"date": date_str, "validated": 0, "mapped": 0}
-        results.append(cumulative_counts)
-        current_date += timedelta(days=1)
-
-    # Populate cumulative_counts with counts from task_history
-    for result in task_history:
-        task_status = result.action_text.split()[5]
-        date_str = result.action_date.strftime("%m/%d")
-        entry = next((entry for entry in results if entry["date"] == date_str), None)
-
-        if entry:
-            if task_status == "VALIDATED":
-                entry["validated"] += 1
-            elif task_status == "MAPPED":
-                entry["mapped"] += 1
-
-    total_validated = 0
-    total_mapped = 0
-
-    for entry in results:
-        total_validated += entry["validated"]
-        total_mapped += entry["mapped"]
-        entry.update({"validated": total_validated, "mapped": total_mapped})
-
-    return results
